Update docs CI
- Move LockedObject from MIOpen to AMDGPU so that other stuff can use
  it.
