+++ .buildkite/pipeline.yml
@@ -136,7 +136,7 @@ steps:
   - label: "Documentation"
     plugins:
       - JuliaCI/julia#v1:
-          version: 1.7
+          version: 1.9-nightly
     command: |
       julia --project -e '
         println("--- :julia: Instantiating project")
